requirements.txt
# Core
django>=4.2,<5.0

# Environment / Config
python-decouple==3.8

# Security — brute-force protection
django-axes==6.4.0

# Production WSGI server
gunicorn==21.2.0

# Static files serving (production)
whitenoise==6.6.0

# Date utilities (calendar-month charts)
python-dateutil==2.9.0

# Pillow (image upload support for future logo fields)
Pillow==10.3.0

# ── Optional — install separately when needed ────────────────────────────────
# PDF invoice export (requires GTK on Windows — install via: pip install weasyprint)
# weasyprint==62.3

# PostgreSQL driver (production only)
# psycopg2-binary==2.9.9
